updating 26 nov eve 6pm

+++ OnlineShopping/OnlineShopping/Models/OrderDetailModel.cs
@@ -9,10 +9,9 @@ namespace OnlineShopping.Models
     {
         public int OrderDetailID { get; set; }
         public DateTime OrderDate { get; set; }
-        public double TotalPrice { get; set; }
+        public int TotalPrice { get; set; }
         public int Quantity { get; set; }
         public int OrderID { get; set; }
         public int ProductID { get; set; }
-       
     }
 }
